[*] move tests to src/test folder
