+++ src/test/client/declarations.js
@@ -8,4 +8,4 @@ const bk = 'WORKBOX UPJERK GORBLE SPECULA SUCKLER FUNNEL INWRAP DOSIS DARNEL CAT
 const pub = 'DCT5dJjvk9k3yTsnJsAph6V8zEPxsAvJ7FCCzqYWiQQyVTiHvReLz';
 const priv = '5JDFQN3T8CFT1ynhgd5s574mTV9UPf9WamkHojBL4NgbhSBDmBj';
 const accountId = '1.2.30';
-dcorejs.dcorejs_lib = dcorejs_lib;
+dcorejs.setLibRef(dcorejs_lib);
